Address review:
* add configure test for _Complex type
* define HAVE_C_COMPLEX
* move workarounds for buggy implementations to Module/_complex.h
