+++ configure
@@ -13845,6 +13845,42 @@ printf "%s\n" "$AIX_BUILDDATE" >&6; }
 	*) ;;
 esac
 
+# check for _Complex C type
+#
+# Note that despite most compilers define __STDC_IEC_559_COMPLEX__ - almost
+# none properly support C11+ Annex G (where imaginary types are mandatory).
+# This is a bug, so we don't rely on presence of __STDC_IEC_559_COMPLEX__.
+if test "$cross_compiling" = yes
+then :
+  ac_cv_c_complex_supported=no
+else $as_nop
+  cat confdefs.h - <<_ACEOF >conftest.$ac_ext
+/* end confdefs.h.  */
+
+#include <complex.h>
+int main(void)
+{
+    double _Complex z = 1 + _Complex_I*2;
+    z = z + _Complex_I*z;
+    return (creal(z) != -1) || (cimag(z) != 3);
+}
+_ACEOF
+if ac_fn_c_try_run "$LINENO"
+then :
+  ac_cv_c_complex_supported=yes
+else $as_nop
+  ac_cv_c_complex_supported=no
+fi
+rm -f core *.core core.conftest.* gmon.out bb.out conftest$ac_exeext \
+  conftest.$ac_objext conftest.beam conftest.$ac_ext
+fi
+
+if test "$ac_cv_c_complex_supported" = "yes"; then
+
+printf "%s\n" "#define HAVE_C_COMPLEX 1" >>confdefs.h
+
+fi
+
 # check for systems that require aligned memory access
 { printf "%s\n" "$as_me:${as_lineno-$LINENO}: checking aligned memory access is required" >&5
 printf %s "checking aligned memory access is required... " >&6; }
