+++ configure.ac
@@ -3763,6 +3763,26 @@ dnl The AIX_BUILDDATE is obtained from the kernel fileset - bos.mp64
 	*) ;;
 esac
 
+# check for _Complex C type
+#
+# Note that despite most compilers define __STDC_IEC_559_COMPLEX__ - almost
+# none properly support C11+ Annex G (where imaginary types are mandatory).
+# This is a bug, so we don't rely on presence of __STDC_IEC_559_COMPLEX__.
+AC_RUN_IFELSE([AC_LANG_SOURCE([[
+#include <complex.h>
+int main(void)
+{
+    double _Complex z = 1 + _Complex_I*2;
+    z = z + _Complex_I*z;
+    return (creal(z) != -1) || (cimag(z) != 3);
+}]])], [ac_cv_c_complex_supported=yes],
+[ac_cv_c_complex_supported=no],
+[ac_cv_c_complex_supported=no])
+if test "$ac_cv_c_complex_supported" = "yes"; then
+    AC_DEFINE([HAVE_C_COMPLEX], [1],
+              [Defined if _Complex C type is available.])
+fi
+
 # check for systems that require aligned memory access
 AC_CACHE_CHECK([aligned memory access is required], [ac_cv_aligned_required],
 [AC_RUN_IFELSE([AC_LANG_SOURCE([[
